Move dependencies to /pkg
